ScheduleData component functions moved over from CronManager, ready to test
