+++ server/globalSettings.js
@@ -1,4 +1,4 @@
-const customPORT = 4323;
+const customPORT = 4324;
 
 
 module.exports = customPORT;
