+++ src/globals.css
@@ -7,3 +7,5 @@
 .nuaFont {
     font-family: 'Aqurine';
 }
+
+
